+++ mcp_tools/password_cracking/hashcat.py
@@ -1,0 +1,36 @@
+# mcp_tools/password_cracking/hashcat.py
+
+from typing import Dict, Any
+
+def register_hashcat_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def hashcat_crack(hash_file: str, hash_type: str, attack_mode: str = "0", wordlist: str = "/usr/share/wordlists/rockyou.txt", mask: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Hashcat for advanced password cracking with enhanced logging.
+
+        Args:
+            hash_file: File containing password hashes
+            hash_type: Hash type number for Hashcat
+            attack_mode: Attack mode (0=dict, 1=combo, 3=mask, etc.)
+            wordlist: Wordlist file for dictionary attacks
+            mask: Mask for mask attacks
+            additional_args: Additional Hashcat arguments
+
+        Returns:
+            Password cracking results
+        """
+        data = {
+            "hash_file": hash_file,
+            "hash_type": hash_type,
+            "attack_mode": attack_mode,
+            "wordlist": wordlist,
+            "mask": mask,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔐 Starting Hashcat attack: mode {attack_mode}")
+        result = hexstrike_client.safe_post("api/tools/hashcat", data)
+        if result.get("success"):
+            logger.info(f"✅ Hashcat attack completed")
+        else:
+            logger.error(f"❌ Hashcat attack failed")
+        return result
