+++ mcp_tools/recon/amass.py
@@ -1,0 +1,30 @@
+# mcp_tools/recon/amass.py
+
+from typing import Dict, Any
+
+def register_amass_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def amass_scan(domain: str, mode: str = "enum", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Amass for subdomain enumeration with enhanced logging.
+
+        Args:
+            domain: The target domain
+            mode: Amass mode (enum, intel, viz)
+            additional_args: Additional Amass arguments
+
+        Returns:
+            Subdomain enumeration results
+        """
+        data = {
+            "domain": domain,
+            "mode": mode,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Amass {mode}: {domain}")
+        result = hexstrike_client.safe_post("api/tools/amass", data)
+        if result.get("success"):
+            logger.info(f"✅ Amass completed for {domain}")
+        else:
+            logger.error(f"❌ Amass failed for {domain}")
+        return result
