--- mcp_tools/web_app_security/additional_security_tools.py
@@ -1,121 +1,0 @@
-# mcp_tools/additional_security_tools.py
-
-from typing import Dict, Any
-
-def register_additional_security_tools(mcp, hexstrike_client, logger):
-
-    @mcp.tool()
-    def amass_scan(domain: str, mode: str = "enum", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Amass for subdomain enumeration with enhanced logging.
-
-        Args:
-            domain: The target domain
-            mode: Amass mode (enum, intel, viz)
-            additional_args: Additional Amass arguments
-
-        Returns:
-            Subdomain enumeration results
-        """
-        data = {
-            "domain": domain,
-            "mode": mode,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Amass {mode}: {domain}")
-        result = hexstrike_client.safe_post("api/tools/amass", data)
-        if result.get("success"):
-            logger.info(f"✅ Amass completed for {domain}")
-        else:
-            logger.error(f"❌ Amass failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def hashcat_crack(hash_file: str, hash_type: str, attack_mode: str = "0", wordlist: str = "/usr/share/wordlists/rockyou.txt", mask: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Hashcat for advanced password cracking with enhanced logging.
-
-        Args:
-            hash_file: File containing password hashes
-            hash_type: Hash type number for Hashcat
-            attack_mode: Attack mode (0=dict, 1=combo, 3=mask, etc.)
-            wordlist: Wordlist file for dictionary attacks
-            mask: Mask for mask attacks
-            additional_args: Additional Hashcat arguments
-
-        Returns:
-            Password cracking results
-        """
-        data = {
-            "hash_file": hash_file,
-            "hash_type": hash_type,
-            "attack_mode": attack_mode,
-            "wordlist": wordlist,
-            "mask": mask,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔐 Starting Hashcat attack: mode {attack_mode}")
-        result = hexstrike_client.safe_post("api/tools/hashcat", data)
-        if result.get("success"):
-            logger.info(f"✅ Hashcat attack completed")
-        else:
-            logger.error(f"❌ Hashcat attack failed")
-        return result
-
-    @mcp.tool()
-    def subfinder_scan(domain: str, silent: bool = True, all_sources: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Subfinder for passive subdomain enumeration with enhanced logging.
-
-        Args:
-            domain: The target domain
-            silent: Run in silent mode
-            all_sources: Use all sources
-            additional_args: Additional Subfinder arguments
-
-        Returns:
-            Passive subdomain enumeration results
-        """
-        data = {
-            "domain": domain,
-            "silent": silent,
-            "all_sources": all_sources,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Subfinder: {domain}")
-        result = hexstrike_client.safe_post("api/tools/subfinder", data)
-        if result.get("success"):
-            logger.info(f"✅ Subfinder completed for {domain}")
-        else:
-            logger.error(f"❌ Subfinder failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def smbmap_scan(target: str, username: str = "", password: str = "", domain: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute SMBMap for SMB share enumeration with enhanced logging.
-
-        Args:
-            target: The target IP address
-            username: Username for authentication
-            password: Password for authentication
-            domain: Domain for authentication
-            additional_args: Additional SMBMap arguments
-
-        Returns:
-            SMB share enumeration results
-        """
-        data = {
-            "target": target,
-            "username": username,
-            "password": password,
-            "domain": domain,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting SMBMap: {target}")
-        result = hexstrike_client.safe_post("api/tools/smbmap", data)
-        if result.get("success"):
-            logger.info(f"✅ SMBMap completed for {target}")
-        else:
-            logger.error(f"❌ SMBMap failed for {target}")
-        return result
